verify state contract for polygon amoy and polygon pos
